move enums to separate files and module
