+++ learning_loop_node/enums/__init__.py
@@ -1,0 +1,6 @@
+from .annotator import AnnotationEventType
+from .detector import OperationMode, OutboxMode, VersionMode
+from .general import CategoryType
+from .trainer import TrainerState
+
+__all__ = ['VersionMode', 'OperationMode', 'OutboxMode', 'AnnotationEventType', 'CategoryType', 'TrainerState']
