+++ learning_loop_node/enums/general.py
@@ -1,0 +1,9 @@
+
+from enum import Enum
+
+
+class CategoryType(str, Enum):
+    Box = 'box'
+    Point = 'point'
+    Segmentation = 'segmentation'
+    Classification = 'classification'
