+++ learning_loop_node/tests/trainer/states/test_state_download_train_model.py
@@ -2,7 +2,7 @@
 import asyncio
 import os
 
-from ....data_classes import TrainerState
+from ....enums import TrainerState
 from ... import test_helper
 from ..state_helper import assert_training_state, create_active_training_file
 from ..testing_trainer_logic import TestingTrainerLogic
