+++ learning_loop_node/tests/trainer/states/test_state_prepare.py
@@ -1,6 +1,7 @@
 import asyncio
 
-from ....data_classes import Context, TrainerState
+from ....data_classes import Context
+from ....enums import TrainerState
 from ....trainer.trainer_logic import TrainerLogic
 from ..state_helper import assert_training_state, create_active_training_file
 from ..testing_trainer_logic import TestingTrainerLogic
