+++ learning_loop_node/tests/trainer/states/test_state_train.py
@@ -1,4 +1,4 @@
-from ....data_classes import TrainerState
+from ....enums import TrainerState
 from ...test_helper import condition
 from ..state_helper import assert_training_state, create_active_training_file
 from ..testing_trainer_logic import TestingTrainerLogic
