+++ learning_loop_node/trainer/trainer_logic.py
@@ -9,7 +9,8 @@ from typing import Coroutine, List, Optional
 
 from dacite import from_dict
 
-from ..data_classes import Detections, ModelInformation, TrainerState, TrainingError
+from ..data_classes import Detections, ModelInformation, TrainingError
+from ..enums import TrainerState
 from ..helpers.misc import create_image_folder, create_project_folder, images_for_ids, is_valid_uuid4
 from .executor import Executor
 from .trainer_logic_generic import TrainerLogicGeneric
